+++ Simulation/TriangleModel.cpp
@@ -65,8 +65,8 @@ void TriangleModel::addPrescribedMotion(const Real startTime, const Real endTime
 
 bool TriangleModel::checkForPrescribedMotion(const Real t, ParticleData &pd)
 {
-	const unsigned int offset = m_indexOffset;
-	const unsigned int nParticles = m_particleMesh.numVertices();
+	const int offset = static_cast<int>(m_indexOffset);
+	const int nParticles = static_cast<int>(m_particleMesh.numVertices());
 
 	bool animated = std::find_if(m_prescribedMotionVector.begin(), m_prescribedMotionVector.end(), 
 									[t] (PrescribedMotion* pm) { return pm->isInTime(t); })
@@ -75,7 +75,7 @@ bool TriangleModel::checkForPrescribedMotion(const Real t, ParticleData &pd)
 	#pragma omp parallel if(nParticles > MIN_PARALLEL_SIZE) default(shared)
 	{
         #pragma omp for schedule(static) 
-		for (unsigned int i = offset; i < offset + nParticles; i++)
+		for (int i = offset; i < offset + nParticles; i++)
 		{
 			if (animated)
 				pd.setParticleState(i, ParticleState::Animated);
@@ -91,8 +91,8 @@ bool TriangleModel::checkForPrescribedMotion(const Real t, ParticleData &pd)
 
 void TriangleModel::applyCurrentPrescribedMotion(const Real t, const Real delta_t, ParticleData& pd)
 {
-	const unsigned int offset = m_indexOffset;
-	const unsigned int nParticles = m_particleMesh.numVertices();
+	const int offset = static_cast<int>(m_indexOffset);
+	const int nParticles = static_cast<int>(m_particleMesh.numVertices());
 
 	// Get current 
 	Real delta = std::numeric_limits<Real>::max();
